feat: separate sections in the homepage into different views
