+++ src/@openarchitex/gatsby-theme-academic-portfolio/components/stat-consulting.js
@@ -1,6 +1,6 @@
 import React from "react"
 import {Box, Grid} from "theme-ui";
-import StatConsultingMDX from "../../sections/stat-consulting";
+import StatConsultingMDX from "../sections/stat-consulting";
 
 const StatConsulting = () => (
     <Grid
@@ -8,7 +8,6 @@ const StatConsulting = () => (
         pb={50}
         pl={4}
         pr={4}
-        id="stat-consulting"
     >
         <Box>
             <StatConsultingMDX />
