+++ src/@openarchitex/gatsby-theme-academic-portfolio/sections/custom-nav-links.mdx
@@ -1,5 +1,6 @@
 import CustomNavLink from "@openarchitex/gatsby-theme-academic-portfolio/src/components/custom-nav-link";
 
+<CustomNavLink link='#about' body='About' />
 <CustomNavLink link='#teaching' body='Teaching' />
 <CustomNavLink link='#research' body='Research' />
 <CustomNavLink link='https://raw.githubusercontent.com/OpenArchitex/TharshannaPortfolio/main/resources/Resume.pdf' body='Resume' newTab={true}/>
